requirements.txt
camelot-py[cv]
PyPDF2<3.0
levenshtein